small edits for UX and UI clarity
